+++ paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/sink/DynamicBucketRowWriteOperator.java
@@ -22,6 +22,9 @@ import org.apache.paimon.data.InternalRow;
 import org.apache.paimon.table.FileStoreTable;
 
 import org.apache.flink.api.java.tuple.Tuple2;
+import org.apache.flink.runtime.io.disk.iomanager.IOManager;
+import org.apache.flink.runtime.state.StateInitializationContext;
+import org.apache.flink.runtime.state.StateSnapshotContext;
 import org.apache.flink.streaming.runtime.streamrecord.StreamRecord;
 
 /**
@@ -31,12 +34,43 @@ public class DynamicBucketRowWriteOperator
         extends TableWriteOperator<Tuple2<InternalRow, Integer>> {
 
     private static final long serialVersionUID = 1L;
+    private transient StoreSinkWriteState state;
 
     public DynamicBucketRowWriteOperator(
             FileStoreTable table,
             StoreSinkWrite.Provider storeSinkWriteProvider,
             String initialCommitUser) {
         super(table, storeSinkWriteProvider, initialCommitUser);
+    }
+
+    @Override
+    public void initializeState(StateInitializationContext context) throws Exception {
+        table = table.copyWithLatestSchema();
+        super.initializeState(context);
+        initStateAndWriter(
+                context,
+                stateFilter,
+                getContainingTask().getEnvironment().getIOManager(),
+                commitUser);
+    }
+
+    @Override
+    protected void initStateAndWriter(
+            StateInitializationContext context,
+            StateValueFilter stateFilter,
+            IOManager ioManager,
+            String commitUser)
+            throws Exception {
+        state = new StoreSinkWriteWithUnionListState(context, stateFilter);
+        write =
+                storeSinkWriteProvider.provide(
+                        table, commitUser, state, ioManager, memoryPool, getMetricGroup());
+    }
+
+    @Override
+    public void snapshotState(StateSnapshotContext context) throws Exception {
+        super.snapshotState(context);
+        state.snapshotState();
     }
 
     @Override
