+++ paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/sink/MultiTablesStoreCompactOperator.java
@@ -103,7 +103,7 @@ public class MultiTablesStoreCompactOperator
                         context, "commit_user_state", String.class, initialCommitUser);
 
         state =
-                new StoreSinkWriteState(
+                new StoreSinkWriteWithUnionListState(
                         context,
                         (tableName, partition, bucket) ->
                                 ChannelComputer.select(
