+++ paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/sink/RowDataStoreWriteOperator.java
@@ -19,218 +19,48 @@
 package org.apache.paimon.flink.sink;
 
 import org.apache.paimon.data.InternalRow;
-import org.apache.paimon.flink.log.LogWriteCallback;
 import org.apache.paimon.table.FileStoreTable;
-import org.apache.paimon.table.sink.SinkRecord;
 
-import org.apache.flink.api.common.functions.RichFunction;
-import org.apache.flink.api.common.functions.util.FunctionUtils;
-import org.apache.flink.api.common.state.CheckpointListener;
-import org.apache.flink.configuration.Configuration;
+import org.apache.flink.runtime.io.disk.iomanager.IOManager;
 import org.apache.flink.runtime.state.StateInitializationContext;
 import org.apache.flink.runtime.state.StateSnapshotContext;
-import org.apache.flink.streaming.api.functions.sink.SinkFunction;
-import org.apache.flink.streaming.api.graph.StreamConfig;
-import org.apache.flink.streaming.api.operators.InternalTimerService;
-import org.apache.flink.streaming.api.operators.Output;
-import org.apache.flink.streaming.api.watermark.Watermark;
-import org.apache.flink.streaming.runtime.streamrecord.StreamRecord;
-import org.apache.flink.streaming.runtime.tasks.ProcessingTimeService;
-import org.apache.flink.streaming.runtime.tasks.StreamTask;
-import org.apache.flink.streaming.util.functions.StreamingFunctionUtils;
 
 import javax.annotation.Nullable;
 
-import java.io.IOException;
-import java.util.List;
-import java.util.Objects;
-
-/** A {@link PrepareCommitOperator} to write {@link InternalRow}. Record schema is fixed. */
-public class RowDataStoreWriteOperator extends TableWriteOperator<InternalRow> {
+/**
+ * A {@link PrepareCommitOperator} to write {@link InternalRow} with StoreSinkWriteState. Record
+ * schema is fixed.
+ */
+public class RowDataStoreWriteOperator extends AbstractRowDataStoreWriteOperator {
 
     private static final long serialVersionUID = 3L;
 
-    @Nullable private final LogSinkFunction logSinkFunction;
-    private transient SimpleContext sinkContext;
-    @Nullable private transient LogWriteCallback logCallback;
-
-    /** We listen to this ourselves because we don't have an {@link InternalTimerService}. */
-    private long currentWatermark = Long.MIN_VALUE;
+    private transient StoreSinkWriteState state;
 
     public RowDataStoreWriteOperator(
             FileStoreTable table,
             @Nullable LogSinkFunction logSinkFunction,
             StoreSinkWrite.Provider storeSinkWriteProvider,
             String initialCommitUser) {
-        super(table, storeSinkWriteProvider, initialCommitUser);
-        this.logSinkFunction = logSinkFunction;
+        super(table, logSinkFunction, storeSinkWriteProvider, initialCommitUser);
     }
 
     @Override
-    public void setup(
-            StreamTask<?, ?> containingTask,
-            StreamConfig config,
-            Output<StreamRecord<Committable>> output) {
-        super.setup(containingTask, config, output);
-        if (logSinkFunction != null) {
-            FunctionUtils.setFunctionRuntimeContext(logSinkFunction, getRuntimeContext());
-        }
-    }
-
-    @Override
-    public void initializeState(StateInitializationContext context) throws Exception {
-        super.initializeState(context);
-
-        if (logSinkFunction != null) {
-            StreamingFunctionUtils.restoreFunctionState(context, logSinkFunction);
-        }
-    }
-
-    @Override
-    protected boolean containLogSystem() {
-        return logSinkFunction != null;
-    }
-
-    @Override
-    public void open() throws Exception {
-        super.open();
-
-        this.sinkContext = new SimpleContext(getProcessingTimeService());
-        if (logSinkFunction != null) {
-            // to stay compatible with Flink 1.18-
-            if (logSinkFunction instanceof RichFunction) {
-                RichFunction richFunction = (RichFunction) logSinkFunction;
-                richFunction.open(new Configuration());
-            }
-
-            logCallback = new LogWriteCallback();
-            logSinkFunction.setWriteCallback(logCallback);
-        }
-    }
-
-    @Override
-    public void processWatermark(Watermark mark) throws Exception {
-        super.processWatermark(mark);
-
-        this.currentWatermark = mark.getTimestamp();
-        if (logSinkFunction != null) {
-            logSinkFunction.writeWatermark(
-                    new org.apache.flink.api.common.eventtime.Watermark(mark.getTimestamp()));
-        }
-    }
-
-    @Override
-    public void processElement(StreamRecord<InternalRow> element) throws Exception {
-        sinkContext.timestamp = element.hasTimestamp() ? element.getTimestamp() : null;
-
-        SinkRecord record;
-        try {
-            record = write.write(element.getValue());
-        } catch (Exception e) {
-            throw new IOException(e);
-        }
-
-        if (record != null && logSinkFunction != null) {
-            // write to log store, need to preserve original pk (which includes partition fields)
-            SinkRecord logRecord = write.toLogRecord(record);
-            logSinkFunction.invoke(logRecord, sinkContext);
-        }
+    protected void initStateAndWriter(
+            StateInitializationContext context,
+            StateValueFilter stateFilter,
+            IOManager ioManager,
+            String commitUser)
+            throws Exception {
+        state = new StoreSinkWriteWithUnionListState(context, stateFilter);
+        write =
+                storeSinkWriteProvider.provide(
+                        table, commitUser, state, ioManager, memoryPool, getMetricGroup());
     }
 
     @Override
     public void snapshotState(StateSnapshotContext context) throws Exception {
         super.snapshotState(context);
-
-        if (logSinkFunction != null) {
-            StreamingFunctionUtils.snapshotFunctionState(
-                    context, getOperatorStateBackend(), logSinkFunction);
-        }
-    }
-
-    @Override
-    public void finish() throws Exception {
-        super.finish();
-
-        if (logSinkFunction != null) {
-            logSinkFunction.finish();
-        }
-    }
-
-    @Override
-    public void close() throws Exception {
-        super.close();
-
-        if (logSinkFunction != null) {
-            FunctionUtils.closeFunction(logSinkFunction);
-        }
-    }
-
-    @Override
-    public void notifyCheckpointComplete(long checkpointId) throws Exception {
-        super.notifyCheckpointComplete(checkpointId);
-
-        if (logSinkFunction instanceof CheckpointListener) {
-            ((CheckpointListener) logSinkFunction).notifyCheckpointComplete(checkpointId);
-        }
-    }
-
-    @Override
-    public void notifyCheckpointAborted(long checkpointId) throws Exception {
-        super.notifyCheckpointAborted(checkpointId);
-
-        if (logSinkFunction instanceof CheckpointListener) {
-            ((CheckpointListener) logSinkFunction).notifyCheckpointAborted(checkpointId);
-        }
-    }
-
-    @Override
-    protected List<Committable> prepareCommit(boolean waitCompaction, long checkpointId)
-            throws IOException {
-        List<Committable> committables = super.prepareCommit(waitCompaction, checkpointId);
-
-        if (logCallback != null) {
-            try {
-                Objects.requireNonNull(logSinkFunction).flush();
-            } catch (Exception e) {
-                throw new IOException(e);
-            }
-            logCallback
-                    .offsets()
-                    .forEach(
-                            (k, v) ->
-                                    committables.add(
-                                            new Committable(
-                                                    checkpointId,
-                                                    Committable.Kind.LOG_OFFSET,
-                                                    new LogOffsetCommittable(k, v))));
-        }
-
-        return committables;
-    }
-
-    private class SimpleContext implements SinkFunction.Context {
-
-        @Nullable private Long timestamp;
-
-        private final ProcessingTimeService processingTimeService;
-
-        public SimpleContext(ProcessingTimeService processingTimeService) {
-            this.processingTimeService = processingTimeService;
-        }
-
-        @Override
-        public long currentProcessingTime() {
-            return processingTimeService.getCurrentProcessingTime();
-        }
-
-        @Override
-        public long currentWatermark() {
-            return currentWatermark;
-        }
-
-        @Override
-        public Long timestamp() {
-            return timestamp;
-        }
+        state.snapshotState();
     }
 }
